refactor: Move grid feature util to shared folder
