+++ lib/project-utility/GridHelper.ts
